+++ src/interpreters/Java_original.rs
@@ -82,13 +82,16 @@ impl Interpreter for Java_original {
     }
 
     fn add_boilerplate(&mut self) -> Result<(), SniprunError> {
-        self.code = String::from(
-            "public class Main {
-            public static void main(String[] args) {
-                ",
-        ) + &self.code
-            + "}
-        }";
+
+        if !Java_original::contains_main(&"public static void main(", &self.code, &"//") {
+            self.code = String::from(
+                "public class Main {
+                public static void main(String[] args) {
+                    ",
+            ) + &self.code
+                + "}
+            }";
+        }
         Ok(())
     }